refactor: service에서 카트를 저장하면 아이디를 반환하는 로직 수정

+++ src/main/java/woowacourse/shoppingcart/dao/CartItemDao.java
@@ -1,5 +1,5 @@
 package woowacourse.shoppingcart.dao;
 
 public interface CartItemDao {
-    Long addCartItem(final Long customerId, final Long productId, int quantity);
+    Long addCartItem(final int customerId, final Long productId, int quantity);
 }
